refactor: modularize MediaWiki API client into separate modules
We need to finish refactoring the MediaWiki API MCP by modularizing the client.py file. The client.py was getting too long and needs to be split into modules in a client_modules directory. Module files should be named like client_page.py, client_search.py, etc. Pay attention to the MediaWikiConfig class that was originally in client.py. MCP tool names must be prefixed with wiki_ and parameter names must align with MediaWiki API spec.

```git-revs
841f758  (Base revision)
05d2bcd  Snapshot before auto-test
e5e926a  Fix test to patch the correct _make_request method on MediaWikiAuthClient
cf72a81  Fix search_pages call to use keyword argument
627884a  Fix import to get MediaWikiConfig from config module
8e8c7e1  Fix import to get MediaWikiConfig from config module
b2de208  Fix import to get MediaWikiConfig from config module
HEAD     Fix import to get MediaWikiConfig from config module
```

codemcp-id: 36-refactor-modularize-mediawiki-api-client-into-sepa
